requirements.txt
absl-py==2.0.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dm-tree==0.1.8
filelock==3.13.1
Flask==3.0.0
fsspec==2023.10.0
gunicorn==21.2.0
h5py==3.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==3.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
Pygments==2.17.2
rich==13.7.0
sympy==1.12
torch==2.1.1
typing_extensions==4.8.0
Werkzeug==3.0.1
